Removed task-slice and added task handling in goal modal
Deleted the task-slice.js file as tasks were made a subpart of goals. Refactored the GoalModal and GoalList to accommodate this change. Added a task creation function, altered the GoalDto on the backend to contain tasks, and updated the goal.state to include tasks handling. Adapted the protected route to remember the user token.